Cleaned some code, working on Hashmap implementation
Cleaned some code : changed variable names, switched to faster small functions, added spaces after colons.

Implementing the necessary structures and functions for the double large prime variation. Work in progress.

+++ include/gaussian_elimination.h
@@ -6,6 +6,6 @@
 #include "structures.h"
 
 void gaussian_elimination(unsigned long relations_len, unsigned long base_size, mpz_t *dense_matrix, mpz_t *res);
-bool row_is_zero(unsigned long relations_len, unsigned long base_size, mpz_t *dense_matrix, unsigned long row_index);
+bool row_is_zero(unsigned long relations_len, unsigned long base_size, mpz_t *dense_matrix, size_t row_index);
 
 #endif // GAUSSIAN_ELIMINATION_H
